Use regex in location to get third party pdf uri
Previously the $request_uri built in Nginx variable
was used to get the third party pdf uri and set the
$upstream. Since ngx.location.capture only supports
internal requests, the uri used here must be to an internal
endpoint. The /id_/ endpoint routes to the third party pdf
uri so this endpoint was re-used for the HEAD request.
In doing so however, that means that the $request_uri
was different depending on whether the request was issued
internally or externally by the client. $request_uri is the uri
of the originating request rather than the current request.
This means it was /http://example.pdf in the internal case
and /id_/http://example.pdf in the external case. Instead of
relying on $request_uri use a regular expression in the location
block to get the third party pdf uri. Note the lua block can also
be removed since it no longer needs to substring $request_uri.

Note merge_slashes has been disabled to that http:// does not
become http:/.

Remove the TODO comment about stripping response Cookies
as it was completed by a previous commit.

Disable access logging for /static files-these just clutter up the log.

Change a comment in the update-pdfjs instructions to refer to
proxy-server rather than via.