Handle redirects in HEAD request for Content-Type
Previously, if the HEAD request for the Content-Type header responded
with a redirect, the Content-Type header from the redirect was used.
Now, a new endpoint is called: /id_/follow_redirect/ which resolves
all redirects to their final location and thus the Content-Type header
that is used is now correct.

A bug was discovered in the /id_/ endpoint while this was being addressed.
When the /id_/ endpoint was simplified to not call a lua script, the
query params were accidentaly dropped from the upstream proxy request.
This issue has been fixed.