Add trailing / to proxied url's with empty paths
Previously if the url did not have a path specified such as:
    http://example.com

It would come back with a 404 response. This has to do with the way
nginx interprets proxy_pass url's with no paths specified. It treats
them as a proxy server and appends the path from the location to the
server. It is our intention in all cases that the proxied url be the full
server + path as opposed to just the server so in cases with very simple
urls we must indicate this to nginx by adding the trailing /.